Moved Person and User to Abstract
