+++ DeliverIT/DeliverIT.csproj
@@ -76,8 +76,8 @@
     <Compile Include="Models\Users\Client.cs" />
     <Compile Include="Models\Users\Courier.cs" />
     <Compile Include="Models\Order.cs" />
-    <Compile Include="Models\Users\Person.cs" />
-    <Compile Include="Models\Users\User.cs" />
+    <Compile Include="Models\Users\Abstract\Person.cs" />
+    <Compile Include="Models\Users\Abstract\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\Dimensions.cs" />
   </ItemGroup>
