+++ DeliverIT/Models/Users/Courier.cs
@@ -1,5 +1,6 @@
 ﻿using DeliverIT.Common.Enums;
 using DeliverIT.Contracts;
+using DeliverIT.Models.Users.Abstract;
 
 namespace DeliverIT.Models.Users
 {
