+++ Tests/Models/OrderTest.cs
@@ -1,11 +1,6 @@
 ﻿using DeliverIT.Contracts;
 using DeliverIT.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace DeliverIT.Tests.Models
 {
